+++ fastly/ngwaf/v1/redactions/api_delete.go
@@ -1,0 +1,43 @@
+package redactions
+
+import (
+	"context"
+	"net/http"
+
+	"github.com/fastly/go-fastly/v10/fastly"
+)
+
+// DeleteInput specifies the information needed for the Delete() function to
+// perform the operation.
+type DeleteInput struct {
+	// Context, if supplied, will be used as the Request's context.
+	Context *context.Context
+	// RedactionID is the redaction identifier (required).
+	RedactionID *string
+	// WorkspaceID is the workspace identifier (required).
+	WorkspaceID *string
+}
+
+// Delete deletes the specified workspace.
+func Delete(c *fastly.Client, i *DeleteInput) error {
+	if i.WorkspaceID == nil {
+		return fastly.ErrMissingWorkspaceID
+	}
+	if i.RedactionID == nil {
+		return fastly.ErrMissingRedactionID
+	}
+
+	path := fastly.ToSafeURL("ngwaf", "v1", "workspaces", *i.WorkspaceID, "redactions", *i.RedactionID)
+
+	resp, err := c.Delete(path, fastly.CreateRequestOptions(i.Context))
+	if err != nil {
+		return err
+	}
+	defer resp.Body.Close()
+
+	if resp.StatusCode != http.StatusNoContent {
+		return fastly.NewHTTPError(resp)
+	}
+
+	return nil
+}