+++ CHANGELOG.md
@@ -9,6 +9,9 @@
 ### Enhancements:
 - feat(ngwaf): add support for virtualpatches ([#680](https://github.com/fastly/go-fastly/pull/680))
 
+- feat(ngwaf/events): add support for the events portion of the ngwaf api ([#684](https://github.com/fastly/go-fastly/pull/684))
+- feat(fastly/ngwaf/v1/redactions): adds CRUD operations for the redactions portion of the ngwaf API ([#685](https://github.com/fastly/go-fastly/pull/685))
+
 ### Bug fixes:
 
 ### Dependencies:
@@ -25,7 +28,6 @@
 ### Bug fixes:
 
 - fix(Domain Discovery): add `omitempty` tag to response types for `suggest` and `status` endpoints.
-
 - fix(fastly/observability_custom_dashboard): add tag to omit context from the json request for observability dashboards([#682](https://github.com/fastly/go-fastly/pull/682))
 
 ## [v10.1.0](https://github.com/fastly/go-fastly/releases/tag/v10.1.0) (2025-05-07)